Move user form to forms
